backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pdf2image==1.17.0
Pillow==10.4.0
PyPDF2==3.0.1
unoconv==0.9.0
requests==2.32.3
pdf2docx==0.5.8 